requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
yt-dlp==2023.11.16
pytube==15.0.0
pydub==0.25.1
ffmpeg-python==0.2.0
pymongo==4.6.1
requests==2.31.0
spotipy==2.23.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
lyricsgenius==3.0.1
python-levenshtein==0.23.0
redis==5.0.1
pydantic==2.5.2
pytgcalls==3.0.0.dev24
pyrogram==2.0.106
TgCrypto==1.2.5
